solve 2.10 - 2.22

+++ src/ua/pp/viktor/chapter01/Exercise11.java
@@ -4,12 +4,12 @@ public class Exercise11 {
 
 	public static void main(String[] args) {
 		int secondsInYear = 365 * 24 * 60 * 60;
-		int populationInYear = secondsInYear/7 - secondsInYear/13 + secondsInYear/45;
-		System.out.println(312032486 + populationInYear);
-		System.out.println(312032486 + 2*populationInYear);
-		System.out.println(312032486 + 3*populationInYear);
-		System.out.println(312032486 + 4*populationInYear);
-		System.out.println(312032486 + 5*populationInYear);
+		double populationInYear = secondsInYear/7.0 - secondsInYear/13.0 + secondsInYear/45.0;
+		System.out.println(312032486 + (int)(populationInYear));
+		System.out.println(312032486 + (int)(2*populationInYear));
+		System.out.println(312032486 + (int)(3*populationInYear));
+		System.out.println(312032486 + (int)(4*populationInYear));
+		System.out.println(312032486 + (int)(5*populationInYear));
 		
 	}
